+++ src/api/v1/mod.rs
@@ -3,5 +3,6 @@ pub mod completions;
 pub mod embeddings;
 pub mod infill;
 pub mod models;
+pub mod props;
 pub mod rerank;
 pub mod tokenize;